Better suggested message #74

+++ movienightbot/util.py
@@ -27,6 +27,9 @@ async def cleanup_messages(
     sec_delay : int
         The number of seconds to wait before deleting the message. Default 10
     """
+    if sec_delay <= 0:
+        # want messages to stay indefinitely so do nothing
+        return
     for message in messages:
         # Adding `delay` kwarg spawns a task, so wrapping that task in a task is redundant...
         # These tasks cause dpytest to break, and py-cord supposedly has "sane rate-limiting"
@@ -47,6 +50,9 @@ async def delete_thread(thread: discord.Thread, sec_delay: int = 10) -> None:
     sec_delay : int
         The number of seconds to wait before deleting the message. Default 10
     """
+    if sec_delay <= 0:
+        # want messages to stay indefinitely so do nothing
+        return
     await asyncio.sleep(sec_delay)
     await thread.delete()
